+++ apollo-api/src/commonMain/kotlin/com/apollographql/apollo3/api/Input.kt
@@ -32,10 +32,10 @@ class Input {
 
 @Deprecated("toInput() is a helper function to help migrating to 3.x " +
     "and will be removed in a future version", ReplaceWith("Optional.presentIfNotNull(this)"))
-@JvmName("toInputOrAbsent")
+@JvmName("-toInputOrAbsent")
 fun <T> T.toInput(): Optional<T> = Optional.presentIfNotNull(this)
 
 @Deprecated("toInput() is a helper function to help migrating to 3.x " +
     "and will be removed in a future version", ReplaceWith("Optional.Present(this)"))
-@JvmName("toInput")
+@JvmName("-toInput")
 fun <T : Any> T.toInput(): Optional<T> = Optional.Present(this)
