+++ apollo-api/src/commonMain/kotlin/com/apollographql/apollo3/api/Operations.kt
@@ -12,7 +12,6 @@ import okio.BufferedSource
 import okio.ByteString
 import kotlin.jvm.JvmName
 
-
 /**
  * Reads a GraphQL Json response like below to a [ApolloResponse]
  * ```
