--- apollo-api/src/jvmMain/kotlin/com/apollographql/apollo3/api/Throws.kt
@@ -1,2 +1,0 @@
-package com.apollographql.apollo3.api
-
