+++ tests/java-tests/src/test/java/test/AdapterTest.java
@@ -1,11 +1,9 @@
 package test;
 
 import com.apollographql.apollo3.api.Adapters;
-import com.apollographql.apollo3.api.CompiledField;
 import com.apollographql.apollo3.api.CustomScalarAdapters;
 import com.apollographql.apollo3.api.NullableAdapter;
 import com.apollographql.apollo3.api.json.BufferedSourceJsonReader;
-import kotlin.jvm.Throws;
 import okio.Okio;
 import org.junit.Test;
